# contracts/openapi/argus-api.yaml
openapi: 3.1.0
info:
  title: ARGUS Platform API
  version: 0.1.0
  description: GENERATED by scripts/contracts.py export from backend/app. Do not hand-edit.
paths:
...
  /api/v1/sites:
    get:
      tags:
      - sites
      summary: List Sites
      operationId: list_sites_api_v1_sites_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_SiteDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - sites
      summary: Create Site
      operationId: create_site_api_v1_sites_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sites/{site_id}:
    get:
      tags:
      - sites
      summary: Get Site
      operationId: get_site_api_v1_sites__site_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          title: Site Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - sites
      summary: Update Site
      operationId: update_site_api_v1_sites__site_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          title: Site Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sites/{site_id}/zones:
    get:
      tags:
      - zones
      summary: List Zones
      operationId: list_zones_api_v1_sites__site_id__zones_get
      security:
      - HTTPBearer: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          title: Site Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ZoneDTO'
                title: Response List Zones Api V1 Sites  Site Id  Zones Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - zones
      summary: Create Zone
      operationId: create_zone_api_v1_sites__site_id__zones_post
      security:
      - HTTPBearer: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          title: Site Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sites/{site_id}/zones/{zone_id}:
    patch:
      tags:
      - zones
      summary: Update Zone
      operationId: update_zone_api_v1_sites__site_id__zones__zone_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          title: Site Id
      - name: zone_id
        in: path
        required: true
        schema:
          type: string
          title: Zone Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/operators:
    get:
      tags:
      - operators
      summary: List Operators
      operationId: list_operators_api_v1_operators_get
      security:
      - HTTPBearer: []
      parameters:
      - name: site_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Site Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_OperatorDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - operators
      summary: Create Operator
      operationId: create_operator_api_v1_operators_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/operators/{operator_id}:
    get:
      tags:
      - operators
      summary: Get Operator
      operationId: get_operator_api_v1_operators__operator_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: operator_id
        in: path
        required: true
        schema:
          type: string
          title: Operator Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - operators
      summary: Update Operator
      operationId: update_operator_api_v1_operators__operator_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: operator_id
        in: path
        required: true
        schema:
          type: string
          title: Operator Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/machines:
    get:
      tags:
      - machines
      summary: List Machines
      operationId: list_machines_api_v1_machines_get
      security:
      - HTTPBearer: []
      parameters:
      - name: site_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Site Id
      - name: machine_type
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - EXCAVATOR
            - DUMPER
            - LOADER
            - DOZER
            - OTHER
            type: string
          - type: 'null'
          title: Machine Type
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_MachineDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - machines
      summary: Create Machine
      operationId: create_machine_api_v1_machines_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachineCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/machines/{machine_id}:
    get:
      tags:
      - machines
      summary: Get Machine
      operationId: get_machine_api_v1_machines__machine_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: machine_id
        in: path
        required: true
        schema:
          type: string
          title: Machine Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - machines
      summary: Update Machine
      operationId: update_machine_api_v1_machines__machine_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: machine_id
        in: path
        required: true
        schema:
          type: string
          title: Machine Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachineUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GeoPoint:
      properties:
        lat:
          type: number
          maximum: 90.0
          minimum: -90.0
          title: Lat
        lon:
          type: number
          maximum: 180.0
          minimum: -180.0
          title: Lon
      type: object
      required:
      - lat
      - lon
      title: GeoPoint
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MachineCreate:
      properties:
        machine_id:
          type: string
          pattern: ^[A-Z0-9_]{2,32}$
          title: Machine Id
        site_id:
          type: string
          title: Site Id
        machine_type:
          type: string
          enum:
          - EXCAVATOR
          - DUMPER
          - LOADER
          - DOZER
          - OTHER
          title: Machine Type
        model:
          type: string
          title: Model
        serial_number:
          type: string
          title: Serial Number
        engine_hours:
          type: number
          minimum: 0.0
          title: Engine Hours
          default: 0.0
      type: object
      required:
      - machine_id
      - site_id
      - machine_type
      - model
      - serial_number
      title: MachineCreate
    MachineDTO:
      properties:
        machine_id:
          type: string
          title: Machine Id
        site_id:
          type: string
          title: Site Id
        machine_type:
          type: string
          enum:
          - EXCAVATOR
          - DUMPER
          - LOADER
          - DOZER
          - OTHER
          title: Machine Type
        model:
          type: string
          title: Model
        serial_number:
          type: string
          title: Serial Number
        engine_hours:
          type: number
          title: Engine Hours
          default: 0.0
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - MAINTENANCE
          - RETIRED
          title: Status
          default: ACTIVE
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - machine_id
      - site_id
      - machine_type
      - model
      - serial_number
      title: MachineDTO
    MachineUpdate:
      properties:
        site_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Site Id
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        engine_hours:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Engine Hours
        status:
          anyOf:
          - type: string
            enum:
            - ACTIVE
            - INACTIVE
            - MAINTENANCE
            - RETIRED
          - type: 'null'
          title: Status
      type: object
      title: MachineUpdate
...
    OperatorCreate:
      properties:
        operator_id:
          type: string
          pattern: ^[A-Z0-9_]{2,32}$
          title: Operator Id
        name:
          type: string
          title: Name
        site_id:
          type: string
          title: Site Id
        experience_level:
          type: string
          enum:
          - JUNIOR
          - INTERMEDIATE
          - SENIOR
          title: Experience Level
          default: JUNIOR
        certification_status:
          type: string
          enum:
          - VALID
          - EXPIRING
          - EXPIRED
          title: Certification Status
          default: VALID
        preferred_language:
          type: string
          title: Preferred Language
          default: en
      type: object
      required:
      - operator_id
      - name
      - site_id
      title: OperatorCreate
    OperatorDTO:
      properties:
        operator_id:
          type: string
          title: Operator Id
        name:
          type: string
          title: Name
        site_id:
          type: string
          title: Site Id
        experience_level:
          type: string
          enum:
          - JUNIOR
          - INTERMEDIATE
          - SENIOR
          title: Experience Level
          default: JUNIOR
        certification_status:
          type: string
          enum:
          - VALID
          - EXPIRING
          - EXPIRED
          title: Certification Status
          default: VALID
        preferred_language:
          type: string
          title: Preferred Language
          default: en
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - MAINTENANCE
          - RETIRED
          title: Status
          default: ACTIVE
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - operator_id
      - name
      - site_id
      title: OperatorDTO
    OperatorUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        site_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Site Id
        experience_level:
          anyOf:
          - type: string
            enum:
            - JUNIOR
            - INTERMEDIATE
            - SENIOR
          - type: 'null'
          title: Experience Level
        certification_status:
          anyOf:
          - type: string
            enum:
            - VALID
            - EXPIRING
            - EXPIRED
          - type: 'null'
          title: Certification Status
        preferred_language:
          anyOf:
          - type: string
          - type: 'null'
          title: Preferred Language
        status:
          anyOf:
          - type: string
            enum:
            - ACTIVE
            - INACTIVE
            - MAINTENANCE
            - RETIRED
          - type: 'null'
          title: Status
      type: object
      title: OperatorUpdate
    Page_MachineDTO_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MachineDTO'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[MachineDTO]
    Page_OperatorDTO_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/OperatorDTO'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[OperatorDTO]
    Page_SiteDTO_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SiteDTO'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[SiteDTO]
...
    SiteCreate:
      properties:
        site_id:
          type: string
          pattern: ^[A-Z0-9_]{2,32}$
          title: Site Id
        name:
          type: string
          title: Name
        location:
          anyOf:
          - $ref: '#/components/schemas/GeoPoint'
          - type: 'null'
        timezone:
          type: string
          title: Timezone
          default: UTC
        configuration:
          additionalProperties: true
          type: object
          title: Configuration
          default: {}
      type: object
      required:
      - site_id
      - name
      title: SiteCreate
    SiteDTO:
      properties:
        site_id:
          type: string
          title: Site Id
        name:
          type: string
          title: Name
        location:
          anyOf:
          - $ref: '#/components/schemas/GeoPoint'
          - type: 'null'
        timezone:
          type: string
          title: Timezone
          default: UTC
        configuration:
          additionalProperties: true
          type: object
          title: Configuration
          default: {}
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - MAINTENANCE
          - RETIRED
          title: Status
          default: ACTIVE
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - site_id
      - name
      title: SiteDTO
...
    SiteUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        location:
          anyOf:
          - $ref: '#/components/schemas/GeoPoint'
          - type: 'null'
        timezone:
          anyOf:
          - type: string
          - type: 'null'
          title: Timezone
        configuration:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Configuration
        status:
          anyOf:
          - type: string
            enum:
            - ACTIVE
            - INACTIVE
            - MAINTENANCE
            - RETIRED
          - type: 'null'
          title: Status
      type: object
      title: SiteUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ZoneCreate:
      properties:
        zone_id:
          type: string
          pattern: ^[A-Z0-9_]{2,32}$
          title: Zone Id
        name:
          type: string
          title: Name
        zone_type:
          type: string
          enum:
          - TASK
          - HAZARD
          - RESTRICTED
          - DISPOSAL
          - SOFT_GROUND
          - OTHER
          title: Zone Type
        geometry:
          additionalProperties: true
          type: object
          title: Geometry
          description: GeoJSON Polygon, coordinates [lon, lat]
      type: object
      required:
      - zone_id
      - name
      - zone_type
      - geometry
      title: ZoneCreate
    ZoneDTO:
      properties:
        zone_id:
          type: string
          title: Zone Id
        site_id:
          type: string
          title: Site Id
        name:
          type: string
          title: Name
        zone_type:
          type: string
          enum:
          - TASK
          - HAZARD
          - RESTRICTED
          - DISPOSAL
          - SOFT_GROUND
          - OTHER
          title: Zone Type
        geometry:
          additionalProperties: true
          type: object
          title: Geometry
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - MAINTENANCE
          - RETIRED
          title: Status
          default: ACTIVE
      type: object
      required:
      - zone_id
      - site_id
      - name
      - zone_type
      - geometry
      title: ZoneDTO
    ZoneUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        zone_type:
          anyOf:
          - type: string
            enum:
            - TASK
            - HAZARD
            - RESTRICTED
            - DISPOSAL
            - SOFT_GROUND
            - OTHER
          - type: 'null'
          title: Zone Type
        geometry:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Geometry
        status:
          anyOf:
          - type: string
            enum:
            - ACTIVE
            - INACTIVE
            - MAINTENANCE
            - RETIRED
          - type: 'null'
          title: Status
      type: object
      title: ZoneUpdate
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer